refactor: moving user API
