+++ routes/register.js
@@ -1,12 +1,8 @@
 const express = require("express");
 const router = express.Router();
 
-const {
-  getRegisterPage,
-  registerUser,
-} = require("../controllers/registerController");
+const { getRegisterPage } = require("../controllers/registerController");
 
 router.route("/").get(getRegisterPage);
-router.route("/api").post(registerUser);
 
 module.exports = router;
